powder 시뮬레이터 입고 개발 완료

+++ mng-infra/src/main/java/kr/co/aim/infra/persistence/db2entity/powder/H2OrderDPEntity.java
@@ -35,7 +35,7 @@ public class H2OrderDPEntity {
     private String pgmMod;
 
     @Column(name = "MODCNT")
-    private Integer modCnt;
+    private Long modCnt;
 
     @Column(name = "CORDERID")
     private String cOrderId;
